add handler for api request, re-use code
