+++ coins/src/main/java/com/ari/coins/framework/data/network/constants/ErrorCode.kt
@@ -1,0 +1,8 @@
+package com.ari.coins.framework.data.network.constants
+
+object ErrorCode {
+    const val UNKNOWN_HOST_EXCEPTION = -1
+    const val CONNECTION_EXCEPTION = -2
+    const val UNKNOWN = -3
+    const val UNKNOWN_LOCAL = -4
+}
